Clinwiki Webpage redesign link
Clinwiki is a nonprofit organization dedicated to helping people find clinical trials they can participate in. As a member of Hack4Impact @ UMich, we worked to redesign their search/results page so that it would be more accessible and user friendly. This is currently a first sketch and we still need to polish it and make it functional.